+++ include/seqan3/range/views/to_upper.hpp
@@ -7,7 +7,7 @@
 
 /*!\file
  * \author Tobias Loka <LokaT AT rki.de>
- * \brief Provides seqan3::views::to_upper.
+ * \brief [DEPRECATED] Provides seqan3::views::to_upper.
  */
 
 #pragma once
@@ -63,13 +63,17 @@ namespace seqan3::views
  * ### Example
  * \include test/snippet/range/views/to_upper.cpp
  * \hideinitializer
+ *
+ * \deprecated Use std::views::transform([](auto && chr){return std::toupper(chr)});
  */
-inline auto const to_upper = deep{std::views::transform([] (auto const in) noexcept
+#ifdef SEQAN3_DEPRECATED_310
+SEQAN3_DEPRECATED_310 inline auto const to_upper = deep{std::views::transform([] (auto const in) noexcept
 {
     static_assert(builtin_character<std::remove_cvref_t<decltype(in)>>,
-                  "The value type of seqan3::views::to_upper must model the seqan3::builtin_character.");
+                 "The value type of seqan3::views::to_upper must model the seqan3::builtin_character.");
     return seqan3::to_upper(in);
 })};
+#endif // SEQAN3_DEPRECATED_310
 
 //!\}